Initial submit of dynamic scene capabilities.
Introduce the concept of a DynamicProperty which stores values for data
over time and allows for the interpolation for a given time.  Create a
DynamicBillboard object out of these properties for describing a billboard
over time.  Introduce DynamicObject class which will aggregate several
primitive types into a single unified object.  Also add a Layer class for
holding multiple objects grouped together.  This is all very much in flux,
and therefore lacks documentation and tests.