Complete re-implementation of JulianDate.fromIso8601
1. Takes into account all formats defined in the ISO8601 standard, including those not supported by Date.parse.
2. Supports leap seconds and sub-millisecond times.
3. Break out isLeapYear into a seperate helper function.
4. Remove toYearFraction as it has problems and is unused.
5. Write a bunch of new tests.

This does NOT add support for more than 4 digit years, as that is a variant of the standard that requires an agreed upon number of digits that we have not settled on yet.  This means that we can't yet support times before 1 B.C. or after 9999 AD.  This will be added in the future.