Add a makeZipFile build target.
Refactor build script to get the "build" target even faster by running steps in parallel,
and only re-write generated files if the content is different.

Separate out the generated file that builds the complete Cesium object from the main.js used
with Almond to simply require the complete object and assign it globally.