TextureAtlas improvements and new Event class
1. Introduce a new utility class Event, for defining event properties.
See documentation for more details.

2. Add three new methods to TextureAtlas to allow for extending an
existing atlas with new textures.  Add an event for when the texture
or textureCoodinates change and make BillboardCollection use it.

With the new Event class, having events throughout the rest of our code
should now be easy and standardized.  I also hope to update any old
even mechanism to use the new standard.

With the TextureAtlas changes, LabelCollection can probably be further
cleaned up and optimized, this will be part of a larger Label cleanup
in the near future.