Add workaround visualizer for CZML cones.
While they should be visualized using ComplexConicSensor and the DynamicConeVisualizer; ComplexConicSensor has major performance issues that need to be addressed.  Instead, I introduced DynamicConeVisualizerUsingCustomSensor.js which is a drop-in replacement for the original visualizer and the new default.  As inferred by the name, this visualizer renders using the CustomSensor isntead.  The visualization is not correct in all cases, but it's better than what we have.  This is a temporary fix until ComplexConicSensor is fixed.