The timeline shouldn't actually set the clock.
Eventually I may remove the clock from it completely, but for now it's at leasat read-only.  It's now up to the app using the timeline to set the time in response to an event.  This gives more control to the developer, who may not want to jump to a time just because the user clicked on it.