Use the fullExtent specified by ESRI service.
It turns out this is the wrong thing to do.  Instead, we should be using
the tileInfo.origin property to identify the northEast corner of the tile
tree, and the tileInfo.lods.resolution times the number of pixels per tile
to determine the actual extent.  So I'm committing this before changing
directions.