Don't pre-load children of visited tiles.
Previously we were loading the child tiles of every tile we visited,
even if the tile wasn't rendered (because it didn't meet SSE
requirements), and all its children were culled.  Now we're only
loading tiles that we want to render.  This may need further refinement,
but it's a definite improvement for now.