Fix issue #148
CustomSensorVolume._computePositions uses a strange algorithm for
computing the vertex positions that I can't quite explain, and that
algorithm generates NaNs when two directions are close together.  The NaN
vertex positions result in missing chunks of the sensor.

The algorithm is apparently trying to compute an appropriate distance
for the sensor vertex positions.  But that shouldn't be tricky - all
vertices at the radial limit are at the same distance, namely, the radial
limit.  So after this change, each direction is simply multiplied by the
radius and that's our vertex position.  It fixes issue #148 and doesn't
have any apparent bad side effects.