Fix jitter.
The u_mode constants used in CentralBodyVS.glsl were wrong, so it was
effectively using the Columbus View vertex transformation even in 3D mode.
Columbus View doesn't do RTC, so that was the cause of the jitter.  Thanks
to @pjcozzi for spotting this after I spent way too long looking at it.