Tile selection improvement.
When computing screen-space error for a tile, use maxError *
cos(maxLatitude) instead of simply maxError.  The cos is a spherical
approximation, so I need something better, but it's good enough for now
and maybe good enough forever for mostly-spherical planets like Earth.
Also improve the imagery cache to avoid allowing multiple copies of a
texture when both are requested at about the same time.