Replace the way beforeAll/afterAll works.
Instead of trying to figure out whether to add blocks to a given spec, add the beforeAll/afterAll blocks
to the suite's queue.  Results don't work perfectly, because the reporter has assumptions that the only thing with
real results is a spec, but an exception in beforeAll/afterAll shows the suite as failed, at least.