Introduce Imagery class.
All tiles that share a texture now share an Imagery instance, and the
state is tracked in that class instead of in TileImagery.  This eliminates
a potential problem where a TileImagery could previously "own" the loading
of a texture, but never actually load it because it ends up out of view.