Finish Ellipsoid implementation, fix major bug, it turns out cartesianToCartographic has been returning the incorrect latitude all of this time.
1. cartesianToCartographic produced incorrect results, this is now fixed.
2. Fixed unit tests which broke after fixing cartesianToCartographic.
3. Full error checking.
4. Static versions of all prototype functions.
5. Functions now provide an optional result parameter for object re-use.
6. 100% unit test code coverage
7. Complete documentation.
8. Updated CHANGES.

This change touched a lot of files, but all of the heavy lifting is in Ellipsoid and EllipsoidSpec.  There's a tone of room for improvement now for code using Ellipsoid to be optmized, but I'll do that with a different pull.  I think I quashed any new issues that cropped up due to the cartesianToCartographic fixed, but if anyone sees anything strange, let me know.