Optimize BoundingSphere.fromPoints
I missed this the first time around, `BoundingSphere.fromPoints` was creating dozens, if not hundreds, of Cartesian3 isntances each frame; depending on the type of data being used.  This was eating up 20% of the CPU when running simple.czml.  I've re-written the function to re-use all scrach variables and everything is much better now.