Improve horizon culling.
Substantially reduce the number of tiles rendered in horizon views by
using the actual ellipsoid for occlusion rather than a spherical
approximation.  The spherical approximation was pretty poor, especially
near the equator, because it was up to 21 km lower than the actual
ellipsoid.

The 'L' view in ImageryTest.html goes from 90 tiles rendered to 77 (14%
less), and the 'C' view goes from 150 tiles to 126 (16% less).

The new horizon test is a bit heavier on the CPU than it needs to be, though.
I'll be optimizing that next.

Thanks to Frank Stoner for all the help working out the math on this.