Add a clearState property to renderer commands.
If the property is set, the command acts as a clear and draw-related
properties are ignored.  A clear command is not required to have a
bounding volume.  If it does, the bounding volume controls the frustums in
which the clear occurs.  If it does not have one, the clear happens in all
frustums but it does NOT force the use of the camera's worst-case near and
far planes.