Separate clear commands from draw commands.
Introduce two new types, `ClearCommand` and `DrawCommand` and remove
`Command`.  Both new types have an `execute` method which is called by
`Scene` to execute the commands without knowing their type.  In both
cases, `execute` just calls the appropriate method (`draw` or `clear`) on
the `Context`.  In the future, we may want to remove knowledge of
commands from the `Context`, in which case the `execute` methods will
become more sophisticated.  `Context.clear` was also changed to take a
`ClearCommand` instead of a `ClearState`, for consistency with
`Context.draw`.  Finally, I changed both `draw` and `clear` to take an
optional `framebuffer` parameter.  This eliminates the need to copy the
commands to change their framebuffer, which can get tricky with multiple
command types.