Cleanup bad DynamicScene tests
There were a handful of tests in DynamicScene that weren't actually testing what they claimed to be.  This was a bad copy/past job on my part back when I did the initial DynamicScene implementation.  All of the tests now actually test the code they are supposed to (though no new bugs were found, which I'll take as a good thing).  This addresses #248