Make sure imagery is ready to render before rendering it.
Specifically, it needs to have a texture and scale computed.  It's
possible for it to be in the ready state but not yet have a texture and
scale, in which case we can't render it yet.  This was causing an
exception in the Imagery Layers Sandcastle app when toggling a layer on
and off quickly.