Eliminate Images folder.
Earth imagery is now in a new package Resources.  Images used by Sandcastle examples are now in the Sandcastle folder.  Images used by the Widgets are now self-contained in the Widgets package.

Also clean up the build file a bit, and create Eclipse launches for all of the build.xml targets.