Fix billboards in 2D and CV.
The camera and billboard coordinate systems are different, for reasons
that aren't entirely clear to me.  For the camera, X is the longitude
direction, Y is the latitude direction, and Z is the height direction.
For the billboard, X is the height direction, Y is the longitude
direction, and Z is the latitude direction.  So I'm swizzling in the VS
are the moment, even though that's probably not the best solution.