VertexArrayFacade purposes.
When specifying attributes for a VertexArrayFacade, it is now possible to
specify a purpose for each.  On commit, a vertex array is created for each
purpose, consisting of all of the attributes tagged for the purpose plus
all attributes for which the purpose is 'all' or unspecified.  The vertex
arrays share as much vertex data as possible.  Effectively, a separate
vertex buffer is created for each purpose+usage.

Also modified BillboardCollection to use this new capability to provide
the `pickColor` attribute only if the pick pass is enabled, and the
`color` attribute only if the color pass is enabled.  This brings the
number of vertex attributes back to 8, which is the number guaranteed by
WebGL.