Replace BingMaps quadkey conversion functions with our own implementation.
This was the only code that came from MSDN sample code, so we no longer need that license information.  The new implementations use bitwise operations to make the actual logic more clear (quadkey digits are base 4 numbers with x and y encoded in each of the 2 bits), and avoid pointless conversions between numeric values and ASCII chars and string comparisons.