More inertial path visualization work
DynamicPositionProperty can now store data in both fixed and inertial reference frames.  Some new functions (currently private) are used by the DynamicPathVisualizer to render in the frame that the data is defined in.  Still needs to be updated to always render earth-fixed for 2D and Columbus view and also handle changing reference frames.