Rework Dojo build process and make it work with the new version.
The CesiumViewer app is now built into its own layer instead of using dojo/dojo.js.  Rename boot.js to CesiumViewerStartup.js and move all code out of there.  It only exists to set up the package config.  Reduce the number of files in the built output by deleting stuff we don't need, and modules that are already included in the app layer.