Produce a built version of CesiumViewerWidget.
The "Cesium Viewer" application is not what people should be using in their own "applications" (simple though those applications may be) that want to use the widget.  So, create a built layer for the CesiumViewerWidget, allowing clients to require that in and recieve all dependencies, allowing them to construct a CesiumViewerWidget instance however they like and put it whereever.