Improve DynamicScene path visualization
Previously, we were only using data points explicitly specified in the CZML to draw paths.  This change makes the sampling algorithm take a maximum step size of 60 seconds so that if points are further apart, we step in-between them.  This allows us to generate certain types of data (such as satellites) with a greter step size in the CZML.  For example, I regenerated simple.czml and it is now 50% smaller with no visual differences.

While the path code is still not in its final form, this change provides a big benefit from a minor tweak.  The easiest way to test this is to use the new simple.czml in master and notice how jagged the orbits are, loading in this branch shows smooth orbits.