Tweak camera tutorial based on comments in #511.
Get rid of the viewport quad creation, instead, just show instructions in HTML.  Add a message indicating you need to click on the widget.  Reorder the code to more closely match the Hello World example this is based on.  Refactor and improve the key handling code so that flags are toggled on and off individually, instead of clearing all flags on keyup, so you can press and release keys independently of each other.