Add dynamic resizing to Cesium Viewer layout
Use media queries to provide 3 different configurations based on screen width.  We still need to manually place the centralBody logo because it's currently not stylable by CSS. This makes the widget look better when embedded or on smaller resolution screens.

Hopefully everyone likes the sizes I picked, but if not, please suggest some new ones.