Fullscreen updates.
1. Move fullscreen button to lower right.
2. Added new SVG images to indicate which action the button will perform.
3. Themed the buttons so they match the rest of viewer.

This isn't quite ready to be merged, but I wanted to get some feedback.  I'm not much for design, so I assume @emackey may want to tweak the SVGs.

Also, I'm considering pulling out this code into it's own widget, but I wasn't sure if that was overkill.  I'm leaning towards doing it (and also de-dojo-fying it in the process).  It will go along nicely with other things we're doing elsewhere (like the cesium-widget branch).