Remove need to call Timeline.updateFromClock.
Register for the `Clock.onTick` event instead, which was introduced with the animation branch.

This is a minor tweak to bring Timeline in line with other widgets (event driven rather than manual updates from the user).  I plan on looking into refactoring Timeline to use knockout/viewModels soon, but I thought this was a good simple change in the mean time.