Fix Two Widgets Sandcastle example
This example exposed an error in the `Event` implementation and cropped
up after the recent Timeline change.  The same functions could not be
registered with multiple scopes (which is what happens if two
instances of the Timeline sharing the same clock).  Even shouldn't really
even be performing this check, so I removed the check and found that remove
was broken too.  I also wrote some new specs to cover this (previously untested) use case.