Fix Polyline crash.
This fixes the following bug.

1. Load simple.czml into Cesium Viewer
2. Go into 2D mode
3. Scrub around with the timeline for a few seconds, you will crash every time.

The issue turned out to be that a polyline had its vertex array computed while it was crossing the IDL (being split into 4 positions from 2).  Later, when it was no longer crossing the IDL, it only had two positions but we didn't bother to recompute the array.  This simply makes the length check happen whenever the positions are dirty and removes a bad check for IDL intersection.