Update billboard buffer usage heuristics.
Apps that use mouse over picking would see performance issues when using a large number of billboards.  This was the result of computeNewBuffersUsage being called during pick, which would result in all usages being set back to STATIC_DRAW and the VertexArrayFacade being recreated every time the mouse moved.  This changes it so we don't call computeNewBuffersUsage on pick passes.

Also, we no longer reset usage properties ever pass, which could result in ping-ponging, instead, we only reset them when we are recreating the vertex for other reasons, such as billbarods being added/removed.