Ignore polylines with less than 2 points.
Previously the check for polyline length happened in `PolylineBucket.prototype.getPolylinePositionsLength`, but this is too late in some circumstance and causes us to index out of bounds of the position array.  This showed up with #706. We now check for length > 1 in sortPolylinesIntoBuckets, which occurs much earlier in the pipeline

I also removed a bunch of unecessary color parameters from the unit tests.