Improvements to CentralBody fragment shader.
Use an if statement instead of the crazy two-step-alpha-zeroing for
excluding a texture outside its extent of applicability.  Avoiding ifs is
sometimes good, but here it just overcomplicates the shader.  In theory,
with dynamic branching, the if could be faster, too.  I wasn't able to
measure any performance difference one way or the other from this change.

Treat alpha like the other properties (hue, gamma, etc.).  Only apply it
if it is not 1.0.

Removing debugging feature of showing texture boundaries.
TileCoordinatesImageryProvider does a better job of that anyway.