Fixes #760 - Globe tiles disappear
For level zero tiles, the bounding sphere tends to be close to the center
of the Earth and not necessarily in a useful direction.  So using it for
the horizon occlusion point produced a pretty crappy point, which was then
considered very slightly not visible when the camera was tipped all the
way down by the pole.  I switched to computing the horizon occlusion point
based on the center of the tile used for RTC instead, which is pretty
close to the same for small tiles, but much better for big ones.