Fix issue #356
This removes the logic from the LabelCollection that attempt to prevent the texture atlas used from growing unbounded. This code negatively impacts performance and can create a hiccup affect during animation. The problem it is attempting to address is most likely not to be encountered in the real world and if we ever run into it, we can revisit this issue.