Material support for DynamicPolyline
Originally, we hardcoded DynamicPolyline to always us a PolylineOutline material type because we did not support general materials in CZML.  Now that the DynamicScene and CZML processing are decoupled, the lack of CZML support for it doesn't matter.  This allows users to use any Material we support in DynamicScene that also supports polylines (currently just ColorMaterialProperty and PolylineOutlineMaterialProperty but others are trivial to add).