Fix SVG path regression with FullscreenButton.
Simply subscribing to a computed observable does not force the computed observable to be evaluated.  If it's never otherwise evaluated, the subscription will never fire.  To work around this, extract the subscribeAndEvaluate helper function from within Animation, and use it everywhere we add a subscriber, to ensure correct behavior no matter what by forcing the evaluation to occur.