Fix Cesium.js generated file.
My recent changes to generate stubs caused the fileset for Cesium.js to be created too early, meaning that shader files could get left out if they were not present on disk before the build started.

Instead of the pathconvert hackery I was using before, I fixed our custom scriptdef tasks to work with file-based ResourceCollections instead of requiring a FileSet, so that we can use the <union> directly.