CZML loading should suspend events
Loading CZML was previously sending events the moment a DynamicObject was created (even though the object had no actual data in it).  It would also send out one event per object, which affects performance.  This change uses the `DynamicObjectCollection.suspectEvents` functionality to send one update with all data loaded at the end.  I thought this was the behavior already, but it wasn't.

I also fixed #1380 while I was in the code.  This was do to all visualizers not cleaning up properly in their destroy.  It was a one line change for each visualizer.