Remove GeometryBatchType
GeometryBatchType was originally a way for GeometryUpdaters to inform the visualizer as to what batch they would ultimately belong in.  Unfortunately, there are just too many potential batches that could be created due to the various combination of properties.  For this reason, the enum was removed and replaced with individual properties on the updater that are used by the visualizer to determine type.