Add changed event for TimeIntervalCollection.
This even is raised whenever the collection of intervals is changed.  This does not attempt to inform the user as to the detials of the changes, just that it happened.  This is from the `dynamicScene-refactor` branch and was extracted out to simplify that PR.

I took the minimal approach here because I didn't want to get wrapped up in the larger TimeInterval, and TimeIntervalCollection refactor that still needs to occur in the future.