Change faceForward defaults
After a bunch of offline discussion, we decided to change the `faceForward` defaults to be `!closed` when the `Appearance` has a `closed` property, and `true` otherwise.  This means that `faceForward` is correctly set by default in every use case we can think of (and I could remove it everwhere we were specifying it).