Fix viewer timeline resize
We size the timeline based on the existence of the animation widget and fullscreen button; however we were not handling the fullscreen button case on resize. If you created the Viewer widget vefore it's container was laid out, the timeline layout would use a width of 0 for the fullscreen button and not update correctly once layout occured.