Fix rendering of debug bounding volumes in 2D and CV
We were using a unit sphere and then deforming it with a model matrix.
This does not work in 2D and CV modes because we do not support setting the
Primitive's modelMatrix property in those modes (because it would require
us to recompute the geometry to handle dateline breaks).

The simple solution is to recreate a primitive every frame, which is
slower but does not matter much when debugging.

Now that we can render these properly, it seems to expose several other issues,
namely that the sphere appear wrong in many cases.  I'm not sure what's going
on and need to look further into this.