Do not check for duplicate transferable objects
Web workers require a unique array of transferable objects.  We never have duplicate items naturally, but did not prevent it from happening.  The performance overhead of this is immense, so we removed the check and put the onus of the geometry creator to copy data that needs to be duplicated.