Improve detection of transferring array buffers.
Remove FeatureDetection.supportsTransferringArrayBuffers which incorrectly used window.postMessage.

Detection is now done internally inside TaskProcessor using a real web worker, which round-trips a typed array to detect a related bug in older Firefox versions.  TaskProcessor sends the result of the detection to the worker along with the task object, instead of having the worker detect itself.

Instead of detecting whether or not ArrayBuffers can be transferred everywhere in Cesium code, we now assume they can, and the TaskProcessor/createTaskProcessorWorker system will internally truncate the array if it's not supported.