Don't refine tiles past our available data.
If a tile's geometry was created by upsampling its parent, and all of its
TileImagery instances are adopted from its parent as well, mark the tile
as UPSAMPLED_ONLY.  When selecting tiles to render, don't refine a parent
tile when all of that tile's children are marked UPSAMPLED_ONLY, because
it's just extra draw commands and GPU work with no benefit.

This reduces the number of tiles rendered in the Half Dome view in the Terrain
Sandcastle example from 243 to 219, which is still way too many.  Other
views are unaffected because they're not close enough to exceed available
data.