Add FeatureDetection.hardwareConcurrency
Chrome Canary recently added `navigator.hardwareConcurrency` which exposes the number of physical cores on the client.  This is a simple change to expose that value (or a default if not defined) on `FeatureDetection`.

I chose 3 as the fallback because according to Steam's hardware survey (http://store.steampowered.com/hwsurvey/) (2 cpus 47.26% 4 cpus 44.94%)

This will be handy when we do #1437 (probably post 1.0), but is useful immediately for `Primitive` creation.