Fix DynamicPathVisualizer handling of ReferenceProperty
`DynamicPathVisualizer` was relying on a hidden `_referenceFrame` property on position objects.  This is not the case with `ReferenceProperty`.  This changes it to use the public API.

Also, the subsampling code for paths was not unwrapping `ReferenceProperty` instances, causing it to fallback to generic sub-optimal sampling.  This addresses that as well and adds tests to verify it.