Fix Camera.pickEllipsoid during morph
Calling `Camera.pickEllipsoid` during a morph would cause the result parameter to be returned, when in reality the result is `undefined` (which is what we already return when a pick is not on the ellipsoid).  I suppose ideally we would actually support pick ellipsoid during a morph, but that's hardly a feature I ever see being useful.

I added a test, but this is also easily testable with the Sandacastle picking example by showing coordinates under a mouse and then morphing.