Refactor Sandcastle to avoid require boilerplate in the code editor.
The code in the editor is automatically surrounded with code to require in the Cesium module.  This removes the need for other buckets except Dojo (for now).
The gallery HTML files can be hand-edited to use a combined Cesium.js file simply by removing the RequireJS scripts from the head and inserting
a script tag for Cesium.js.
Various other cleanup in the Sandcastle code itself.  Gallery files are reloaded from the server when clicking them now, previously the contents would be cached in memory.