Fix Sandcastle regression
`Sandcastle.addToolbarMenu` will make the first option a default if it has an `onselect` function and a default action is not already set.

This fixes a problem with the billboards and label examples where they were manually calling the first command and then it would get reset at the end of loading.